infra: blue/green deployments
