+++ apps/infra/Pulumi.app-production-web.yaml
@@ -1,0 +1,1 @@
+encryptionsalt: v1:H8lEi2UXZLE=:v1:JKwRl5r98f0cySKx:7XXIpIHQtXfXtN5UyhorIAyb17mjhg==
