--- apps/infra/Pulumi.web-production.yaml
@@ -1,1 +1,0 @@
-encryptionsalt: v1:IJ9QRxpLtPE=:v1:8p21+tCF/173fNCd:TsqDebvtbxTZODi0KaTCth5HUaS/+Q==
