+++ apps/infra/src/app/production/workers.ts
@@ -8,22 +8,17 @@ import {
   ecsTaskExecutionRole,
   privateSubnets,
   resolve,
-  vpcId,
-} from '../shared'
+} from '../../shared'
 import { coreStack, environment } from './shared'
 
-const DNS_ADDRESS = 'gateway.latitude.so'
+const repo = new aws.ecr.Repository('latitude-llm-workers-repo')
 
-// Create an ECR repository
-const repo = new aws.ecr.Repository('latitude-llm-gateway-repo')
-
-// Build and push the Docker image
 const token = await aws.ecr.getAuthorizationToken()
-const image = new docker.Image('LatitudeLLMGatewayImage', {
+const image = new docker.Image('LatitudeLLMWorkersImage', {
   build: {
     platform: 'linux/amd64',
     context: resolve('../../../'),
-    dockerfile: resolve('../../../apps/gateway/docker/Dockerfile'),
+    dockerfile: resolve('../../../apps/workers/docker/Dockerfile'),
     cacheFrom: {
       images: [pulumi.interpolate`${repo.repositoryUrl}:latest`],
     },
@@ -36,11 +31,9 @@ const image = new docker.Image('LatitudeLLMGatewayImage', {
   },
 })
 
-// Create a Fargate task definition
-const containerName = 'LatitudeLLMGatewayContainer'
-// Create the log group
-const logGroup = new aws.cloudwatch.LogGroup('LatitudeLLMGatewayLogGroup', {
-  name: '/ecs/LatitudeLLMGateway',
+const containerName = 'LatitudeLLMWorkersContainer'
+const logGroup = new aws.cloudwatch.LogGroup('LatitudeLLMWorkersLogGroup', {
+  name: '/ecs/LatitudeLLMWorkers',
   retentionInDays: 7,
 })
 
@@ -48,7 +41,7 @@ const taskDefinition = pulumi
   .all([logGroup.name, image.imageName, environment])
   .apply(
     ([logGroupName, imageName, environment]) =>
-      new aws.ecs.TaskDefinition('LatitudeLLMGatewayTaskDefinition', {
+      new aws.ecs.TaskDefinition('LatitudeLLMWorkersTaskDefinition', {
         family: 'LatitudeLLMTaskFamily',
         cpu: '256',
         memory: '512',
@@ -61,16 +54,13 @@ const taskDefinition = pulumi
             name: containerName,
             image: imageName,
             essential: true,
-            portMappings: [
-              { containerPort: 8080, hostPort: 8080, protocol: 'tcp' },
-            ],
             environment,
             healthCheck: {
               command: [
                 'CMD-SHELL',
-                'curl -f http://localhost:8080/health || exit 1',
+                'curl -f http://localhost:3002/health || exit 1',
               ],
-              interval: 10,
+              interval: 30,
               timeout: 5,
               retries: 3,
               startPeriod: 60,
@@ -88,42 +78,8 @@ const taskDefinition = pulumi
       }),
   )
 
-const targetGroup = new aws.lb.TargetGroup('LatitudeLLMGatewayTg', {
-  port: 8080,
-  vpcId,
-  protocol: 'HTTP',
-  targetType: 'ip',
-  healthCheck: {
-    path: '/health',
-    interval: 5,
-    timeout: 2,
-    healthyThreshold: 2,
-    unhealthyThreshold: 2,
-  },
-  deregistrationDelay: 5,
-})
-
-const defaultListenerArn = coreStack.requireOutput('defaultListenerArn')
-
-new aws.lb.ListenerRule('LatitudeLLMGatewayListenerRule', {
-  listenerArn: defaultListenerArn,
-  actions: [
-    {
-      type: 'forward',
-      targetGroupArn: targetGroup.arn,
-    },
-  ],
-  conditions: [
-    {
-      hostHeader: {
-        values: [DNS_ADDRESS],
-      },
-    },
-  ],
-})
-
 const cluster = coreStack.requireOutput('cluster') as pulumi.Output<Cluster>
-new aws.ecs.Service('LatitudeLLMGateway', {
+export const service = new aws.ecs.Service('LatitudeLLMWorkers', {
   cluster: cluster.arn,
   taskDefinition: taskDefinition.arn,
   desiredCount: 2,
@@ -135,19 +91,10 @@ new aws.ecs.Service('LatitudeLLMGateway', {
     assignPublicIp: false,
     securityGroups: [ecsSecurityGroup],
   },
-  loadBalancers: [
-    {
-      targetGroupArn: targetGroup.arn,
-      containerName,
-      containerPort: 8080,
-    },
-  ],
   tags: {
-    digest: image.repoDigest,
+    diggest: image.repoDigest,
   },
   triggers: {
     diggest: image.repoDigest,
   },
 })
-
-export const serviceUrl = pulumi.interpolate`https://${DNS_ADDRESS}`
